requirements.txt
Scrapy==2.11.1
scrapy-redis==0.7.3
scrapy-rabbitmq-link==0.5.0
psycopg2-binary==2.9.9
w3lib==2.1.2
beautifulsoup4==4.12.3